requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
graphviz>=0.20.0
pygraphviz>=1.11
plotly>=5.15.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
ydata-profiling>=4.5.0
openpyxl>=3.1.0
xlrd>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
